Split out a GameData base class for Event.
